Rename reset password files
